+++ packages/react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @caderno-ui/react
 
+## 0.4.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @caderno-ui/elements@0.4.1
+
 ## 0.4.0
 
 ### Minor Changes